Move HTTP/1.x parsing code in Frame to new subclass Http1Frame.
