+++ src/Kestrel.Core/Internal/Http/FrameOfT.cs
@@ -11,7 +11,7 @@ using Microsoft.Extensions.Logging;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
 {
-    public class Frame<TContext> : Frame
+    public class Frame<TContext> : Http1Frame
     {
         private readonly IHttpApplication<TContext> _application;
